refactor(book scraper): OOP end of implementation
